moved all classes to a "classes" folder
