+++ addon/classes/beat-track.js
