+++ addon/classes/beat.js
