+++ addon/classes/note.js
