+++ addon/classes/object-like-map.js
