+++ app/classes/object-like-map.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-audio/classes/object-like-map';
