+++ app/classes/track.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-audio/classes/track';
