--- app/utils/beat-track.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-audio/utils/beat-track';
