--- app/utils/get-prop-from-sound.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-audio/utils/get-prop-from-sound';
