--- app/utils/note.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-audio/utils/note';
