--- app/utils/object-like-map.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-audio/utils/object-like-map';
